+++ ChangeLog.md
@@ -4,6 +4,7 @@ IMPROVEMENTS
 - added support for `use_embedded_backup_restore: true` with empty `embedded_backup_disk` value, tested on S3/GCS over S3/AzureBlobStorage, fix [695](https://github.com/Altinity/clickhouse-backup/issues/695)
 - `--rbac, --rbac-only, --configs, --configs-only` now works with `use_embedded_backup_restore: true`
 -- `--data` for `restore` with `use_embedded_backup_restore: true` will use `allow_non_empty_tables=true` to allow fix [756](https://github.com/Altinity/clickhouse-backup/issues/756)
+- added `--diff-from-remote` parameter for `create` command, will copy only new data parts object disk data, also allows to download properly object disk data from required backup during `restore`, fix [865](https://github.com/Altinity/clickhouse-backup/issues/865)
  
 BUG FIXES
 - continue `S3_MAX_PARTS_COUNT` default value from `2000` to `4000` to continue decrease memory usage for S3 
@@ -11,6 +12,7 @@ BUG FIXES
 - restore SQL UDF functions after restore tables
 - execute `ALTER TABLE ... DROP PARTITION` instead of `DROP TABLE` for `restore` and `restore_remote` with parameters `--data --partitions=...`, fix [756](https://github.com/Altinity/clickhouse-backup/issues/756) 
 - fix wrong behavior for `freeze_by_part` + `freeze_by_part_where`, fix [855](https://github.com/Altinity/clickhouse-backup/issues/855)
+- apply `CLICKHOUSE_SKIP_TABLES_ENGINES` during `create` command
 
 # v2.4.34
 BUG FIXES